Optimized e and v shortcut mechanism
It's far better to determine the editor upon start than to do it every
time anew. Further, this way it's more flexible and organized.

The e command is my cross-machine way of how I like files to be opened
to be edited or viewed. Of course, this doubles the open command to some
degree; and it isn't as reliable as working with mime types.

Well, for now, this is my best take on the topic.