Returned to using aliases
Because functions can be recursive while alias are not, I decided
to return to using aliases for aliases instead of functions. Of course,
it would be possible to determine if the call would be recursive and
use "command", but that would make things more complicated. It's always
better to stay as simple as possible.

So, I'm using functions only where it's required, namely in aliasing
OS X applications for use with smart_open.